refactors handlers
